Turn 'save' anchor into button
The click event for the 'save' returned 'false' to prevent the anchor from navigating to another page. Webpack discarded this during compilation because it saw no need for it. To prevent this, 'Event.preventDefault()' and 'Event.stopPropagation()' could be called manually. Instead, the anchor can also be replaced with a button to prevent the browser from navigating to the destination.